+++ buildSrc/src/main/java/Config.kt
@@ -44,10 +44,6 @@ object Config {
         val timber = "com.jakewharton.timber:timber:4.7.1"
         val okhttpBom = "com.squareup.okhttp3:okhttp-bom:$okHttpVersion"
         val okhttp = "com.squareup.okhttp3:okhttp"
-        // only bump gson if https://github.com/google/gson/issues/1597 is fixed
-        private val gsonVersion = "2.8.5"
-        val gsonDep = "com.google.code.gson:gson"
-        val gson = "$gsonDep:$gsonVersion"
         val leakCanary = "com.squareup.leakcanary:leakcanary-android:2.7"
 
         private val lifecycleVersion = "2.2.0"
@@ -119,7 +115,6 @@ object Config {
         val awaitility = "org.awaitility:awaitility-kotlin:4.1.0"
         val mockWebserver = "com.squareup.okhttp3:mockwebserver:${Libs.okHttpVersion}"
         val mockWebserver3 = "com.squareup.okhttp3:mockwebserver:3.14.9"
-        val jsonUnit = "net.javacrumbs.json-unit:json-unit:2.28.0"
     }
 
     object QualityPlugins {